refactor: view, controller, model로 패키지 관리
