+++ src/main/java/lotto/Application.java
@@ -2,7 +2,7 @@ package lotto;
 
 public class Application {
     public static void main(String[] args) {
-        LottoController controller = new LottoController();
+        controller.LottoController controller = new controller.LottoController();
         controller.run();
     }
 }
