+++ src/main/java/lotto/Model/Bonus.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.Model;
+
+import lotto.Model.enums.ErrorMessage;
 
 import java.util.List;
