+++ src/main/java/lotto/Model/Draw.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.Model;
 
 import java.util.Arrays;
 import java.util.List;
