+++ src/main/java/lotto/View/Validation.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.View;
+
+import lotto.Model.enums.ErrorMessage;
 
 public class Validation {
